requirements.txt
# requirements.txt
Flask==2.3.3
python-dateutil==2.8.2
requests==2.31.0